main process - refactor handlers and libs Fixes #60
