+++ src/state.ts
@@ -1,4 +1,9 @@
 import { IApplicationState } from "./types";
-export const applicationState : IApplicationState = {
+
+export const applicationState: IApplicationState = {
   pageName: ``
 };
+
+export function statePageUpdateHandler(event: Event, pageName: string):void {
+  applicationState.pageName = pageName;
+}
